Refactors common constants, types and colors
