+++ src/components/app/App.module.css
@@ -1,3 +1,5 @@
+@value feltGreen from "../common/colors.css";
+
 @value spacing: 1rem;
 
 .felt,
@@ -6,7 +8,7 @@
 }
 
 .felt {
-  background-color: #006e3b;
+  background-color: feltGreen;
   box-shadow: inset 0 0 10rem black;
   height: calc(100vh - 2 * spacing);
   padding: spacing;
