+++ src/components/common/colors.css
@@ -1,0 +1,6 @@
+@value blackBoard: #282828;
+@value chalkWhite: #fefeff;
+@value feltGreen: #006e3b;
+@value playingCardBlack: #000;
+@value playingCardRed: #ff534f;
+@value playingCardWhite: #fff;
