+++ src/components/common/constants.ts
@@ -1,0 +1,5 @@
+import { Team } from "./types";
+
+export const MAXIMUM_POINTS = 14;
+export const MINIMUM_POINTS = 6;
+export const TEAMS: Team[] = ["us", "they"];
