+++ src/components/keypad/types.ts
@@ -1,4 +1,4 @@
-import { Score } from "components/common";
+import { Score } from "components/common/types";
 
 export interface Props {
   updateRounds: (score: Score) => void;
