+++ src/components/toolbar/index.ts
@@ -1,0 +1,1 @@
+export { Toolbar } from "./Toolbar";
